The great update of 2025.

+++ src/load_balanced/make_DT_knot.c
@@ -18,13 +18,17 @@
  ******************************************************************************/
 #include "load_balanced.h"
 
- /*Function which makes a DT knot given the DT code as a string*/
-struct DT_knot* make_DT_knot(const char* const DT_code) {
-	/* Length of the string is the number of crossings, and the string is copied to the knot */
-	if (DT_code == NULL) return NULL;
-	struct DT_knot* temp = safe_malloc(sizeof(struct DT_knot));
-	temp->number_of_crossings = (int)strlen(DT_code);
-	temp->DT_code = safe_malloc((size_t) (temp->number_of_crossings + 1));
-	strcpy(temp->DT_code, DT_code);
-	return temp;
+/*  Function which makes a DT knot given the DT code as a string              */
+struct DT_knot* make_DT_knot(const char * const DT_code)
+{
+    /* Length of the string is the number of crossings, and the string is     *
+     * copied to the knot.                                                    */
+    if (DT_code == NULL)
+	    return NULL;
+
+    struct DT_knot* temp = safe_malloc(sizeof(struct DT_knot));
+    temp->number_of_crossings = strlen(DT_code);
+    temp->DT_code = safe_malloc(temp->number_of_crossings + 1);
+    strcpy(temp->DT_code, DT_code);
+    return temp;
 }
