+++ src/load_balanced/make_link.c
@@ -20,7 +20,10 @@
 #include "load_balanced.h"
 
 /* Function to make a link given its number of components, number of crossings in each component, and first crossing in each component */
-struct link* make_link(const int number_of_components, int* const number_of_crossings_in_components, struct crossing** const first_crossing_in_components) 
+struct link*
+make_link(const size_t number_of_components,
+          size_t * const number_of_crossings_in_components,
+		  struct crossing** const first_crossing_in_components)
 {
 	struct link* temp = (struct link*)safe_malloc(sizeof(struct link));
 	temp->number_of_components = number_of_components;
